==> rewriting/data_creation/movie/1_get_sentences_for_characters.ipynb <==
import os
# nvidia-smi
os.environ['CUDA_VISIBLE_DEVICES'] = '0'

import pandas as pd
import csv
import numpy as np
from tqdm import tqdm
from itertools import groupby
from names_dataset import NameDataset
tqdm.pandas()

import warnings
warnings.filterwarnings("ignore")

PREFIX = "../data/"
PA_PATH = PREFIX + "sap2017-connotation-frames-power-agency/"
MS_PATH = PREFIX + "bamman2013-movie-summaries/MovieSummaries/"
NLP_PATH = MS_PATH + 'corenlp_plot_summaries/'

import sys
from mov_nlp_utils import *

# read data

@inproceedings{bamman-etal-2013-learning,
    title = "Learning Latent Personas of Film Characters",
    author = "Bamman, David  and
      O{'}Connor, Brendan  and
      Smith, Noah A.",
    booktitle = "Proceedings of the 51st Annual Meeting of the Association for Computational Linguistics (Volume 1: Long Papers)",
    month = aug,
    year = "2013",
    address = "Sofia, Bulgaria",
    publisher = "Association for Computational Linguistics",
    url = "https://aclanthology.org/P13-1035",
    pages = "352--361",
}

Get data at http://www.cs.cmu.edu/~ark/personas/

sums = pd.read_csv(MS_PATH + 'plot_summaries.txt', sep='\t', names=['id', 'text'])
sums = sums.set_index('id')
sums.head()

cha = pd.read_csv(MS_PATH + 'character.metadata.tsv', sep='\t', names=['id', 'fid', 'date', 'cha_name', 'actor_birth', 'actor_gender', 'actor_height', 'actor_ethnicity', 'actor_name', 'actor_age', 'cha_actor_fid', 'cha_fid', 'actor_fid'])
cha = cha.drop(columns=['actor_height', 'actor_ethnicity', 'cha_actor_fid'])
cha.index = pd.MultiIndex.from_arrays(cha[['id', 'cha_fid']].values.T, names=['mid', 'cid'])

# character must have a name and gender (TODO: gender name lists?)
cha = cha[cha.cha_name.notna() & cha.actor_gender.notna()] 

# not unique index when multiple actors played one role -> keep only the first
cha = cha[~cha.index.duplicated('first')] 

# remove chars without summary
cha = cha[cha['id'].isin(sums.index)]

## preprocess names

name_dataset = NameDataset()

# new colum cha_names
cha['cha_names'] = cha['cha_name'].str.split(' ')

# filter names with only one letter
cha = cha[cha.cha_name.apply(len) > 2]

def get_firstname(names):
    max_score = 0
    first_name = ''
    for i in range(len(names)):
        curr_score = name_dataset.search_first_name(names[i]) 
        if curr_score > max_score:
            max_score = curr_score
            first_name = names[i]
            
    # threshold
    if max_score > 0.01:
        return first_name
    else:
        return None

cha['firstname'] = cha['cha_names'].apply(get_firstname)

cha = cha[cha.firstname.notna()]

#remove movie summaries without characters
sums = sums[sums.index.isin(cha['id'])]

len(cha), len(sums)

# get sentences for characters

def get_ner_persons_per_sentence(mid, sents):    
    persons = [] # list with person token list for each sentence
    person_tokens = [] # list of lists with subsequent person tokens
    for s in sents['sentence']:
        person_token = [] # list of subsequent person tokens
        try:
            for t in s['tokens']['token']:
                if isinstance(t, dict) and t['NER'] == 'PERSON': # is person
                    person_token.append((t['word'], (s['@id'], t['@id'])))
                else:
                    if len(person_token) > 0:
                        person_tokens.append(person_token)
                    person_token = []
        except:
            pass # parsing error
    return person_tokens

def get_best_character_match(mid, person_tokens):
    chars = cha[cha['id'] == mid]
    result = []
    
    for token_list in person_tokens:
        cid = None
        name = " ".join([t[0] for t in token_list]) # merge to full name       
        if len(chars[chars.cha_name == name]) == 1: # no confusions
            # full name found in text
            cid = chars[chars.cha_name == name].index.item()[1]

        if not cid:
            # maybe it is only the first name?            
            if len(token_list) == 1:
                if len(chars[chars.firstname == name]) == 1: # no confusions
                    # first name found in text
                    cid = chars[chars.firstname == name].index.item()[1]
                                                                      
        if cid:
            result.append({
                'name': name,
                'cid': cid,
                'st_ids' : [token[1] for token in token_list],
            })
    return result

def group_by_character(character_matches):    
    result = []
    character_matches.sort(key=lambda x: x['cid'])
    for k, v in groupby(character_matches, key = lambda x: x['cid']):
        values = [value for value in list(v)]
        result.append({
            'cid': k, 
            'names': list(set([entry['name'] for entry in values])),
            'st_ids': [entry['st_ids'] for entry in values]
        })
    return result

def coref_to_st_id_list(coref):
    result = []
    try:
        for i in range(len(coref['coreference'])):
            coref_elem = coref['coreference'][i]
            coref_group = []

            for mention in coref_elem['mention']:
                mention_elems = []
                for i in range(int(mention['start']), int(mention['end'])): # end is exclusive
                    st_id = (mention['sentence'], str(i))
                    mention_elems.append(st_id)
                coref_group.append(mention_elems)
            result.append(coref_group)
    except:
        pass # parsing error
    return result

def find_coreferences(mid, character_occurences, coref):
    result = []
    
    for char_occur in character_occurences:
        corefs = []
        char_st_ids = [item for sublist in char_occur['st_ids'] for item in sublist] # flatten char occur st_ids
        coref_st_ids = coref_to_st_id_list(coref) # get corefs as nested st_id lists
        
        for coref_group in coref_st_ids:
            coref_group_flat = [item for sublist in coref_group for item in sublist] # flatten coref_groups 
            intersection = set(tuple(x) for x in char_st_ids).intersection(set(tuple(x) for x in coref_group_flat))
            if len(intersection) > 0 and not coref_group in char_occur['st_ids']:
                corefs.append(coref_group)
                
        char_occur['coref'] = corefs
        result.append(char_occur)

    return result

def get_sentences_for_characters(characters_w_coreferences, sents):
    result = {}
    
    for character in characters_w_coreferences:
        elem = character
        sentences = []
        sent_ids = []
        
        # sentences for st_ids
        for st_id in character['st_ids']:
            sent_id = st_id[0][0]
            
            if not sent_id in sent_ids:
                sent_ids.append(sent_id)
                sentences.append((get_sentence(sents, sent_id), st_id)) # only sentence interesting, take s_id from first elem
            
        
        # sentences for corefs
        for coref_group in character['coref']:
            for mention in coref_group:
                sent_id = mention[0][0]
                
                if not sent_id in sent_ids:
                    sent_ids.append(sent_id)
                    sentences.append((get_sentence(sents, sent_id), mention)) # only sentence interesting, take s_id from first elem
             
                
        elem['sentences'] = sentences
        result[(character['cid'], mid)] = elem
        
    return result

# put all together

def get_sentences_pipeline(mid):
    sents, coref = get_nlp_file(mid)
    
    ner_persons = get_ner_persons_per_sentence(mid, sents)
    character_matches = get_best_character_match(mid, ner_persons)
    character_occurences = group_by_character(character_matches)
    characters_w_coreferences = find_coreferences(mid, character_occurences, coref)
    return get_sentences_for_characters(characters_w_coreferences, sents)

mid = 31186339 # (just for development)

result = get_sentences_pipeline(mid)

result = {}
for mid in tqdm(sums.index, total=len(sums)):
    next_result = get_sentences_pipeline(mid)
    
    key_intersect = result.keys() & next_result.keys() 
    if (len(key_intersect)) == 0:
        result.update(next_result)
    else:
        print(mid)

df = pd.DataFrame.from_dict(result, orient='index')

df

df.sentences.apply(len).value_counts()

df.to_pickle(MS_PATH + 'character_sentences.pkl')
