==> rewriting/data_creation/movie/2_verb_prep_for_sentences.ipynb <==
import os
# nvidia-smi
os.environ['CUDA_VISIBLE_DEVICES'] = '0'

import pandas as pd
import csv
import numpy as np
from tqdm import tqdm
tqdm.pandas()

import warnings
warnings.filterwarnings("ignore")

PREFIX = "../data/"
PA_PATH = PREFIX + "sap2017-connotation-frames-power-agency/"
MS_PATH = PREFIX + "bamman2013-movie-summaries/MovieSummaries/"
NLP_PATH = MS_PATH + 'corenlp_plot_summaries/'

import sys
from mov_nlp_utils import *

# read data

%%time
df_raw = pd.read_pickle(MS_PATH + 'character_sentences.pkl')

df = df_raw.explode('sentences') 

len(df), len(df_raw)

df['mention'] = df.sentences.apply(lambda s: s[1])

df['sentence'] = df.sentences.apply(lambda s: s[0])

df.drop(['st_ids', 'coref', 'sentences'], axis=1, inplace=True)

df.head()

df[df.mention.apply(lambda x: True if len(x) > 1 else False)]

def get_verb_for_character(row):     
    verbs = []
    result = []
    
    
    if isinstance(row.sentence, dict) and row.sentence['collapsed-dependencies']:
        # verbs
        verbs = [get_token_for_sentence(row.sentence, t['governor']['@idx']) for t in row.sentence['collapsed-dependencies']['dep']
                      if isinstance(t, dict) and t['@type'] in ['nsubj']
                      if 'VB' in get_token_for_sentence(row.sentence, t['governor']['@idx'])['POS']
                      and (row.sentence['@id'], t['dependent']['@idx']) in row.mention]
    
        # prepositions
        for verb in verbs:
            prep = [get_token_for_sentence(row.sentence, t['dependent']['@idx']) for t in row.sentence['collapsed-dependencies']['dep']
                      if isinstance(t, dict) and t['@type'] == 'prep'
                      if get_token_for_sentence(row.sentence, t['dependent']['@idx'])['POS'] in ['IN', 'TO']
                      and get_token_for_sentence(row.sentence, t['governor']['@idx']) == verb]
            
            if len(prep) > 0:
                result.append((verb, prep))
            else:
                result.append((verb)) 
                            
    return result

df['verbs'] = df.progress_apply(get_verb_for_character, axis=1)

df.verbs.apply(len).value_counts()

df = df[df.verbs.apply(len) > 0]

df.verbs.apply(len).value_counts()

df.verbs

df.to_pickle(MS_PATH + 'character_sentences_verbs.pkl')
