==> rewriting/data_creation/movie/3_final_csv.ipynb <==
import os
# nvidia-smi
os.environ['CUDA_VISIBLE_DEVICES'] = '0'

import pandas as pd
import csv
import numpy as np
from tqdm import tqdm
tqdm.pandas()

import warnings
warnings.filterwarnings("ignore")

PREFIX = "../data/"
PA_PATH = PREFIX + "sap2017-connotation-frames-power-agency/"
MS_PATH = PREFIX + "bamman2013-movie-summaries/MovieSummaries/"
NLP_PATH = MS_PATH + 'corenlp_plot_summaries/'

import sys
from mov_nlp_utils import *

# read data

## sentences

MS_PATH + 'character_sentences_verbs.pkl'

%%time
df = pd.read_pickle(MS_PATH + 'character_sentences_verbs.pkl')

df['verb'] = df.verbs.apply(lambda v: v[0]) # first verb only (highest in dep parse tree, most important)

df.drop('verbs', inplace=True, axis=1)

df.head()

len(set(df.index)) # 32059 different characters

## power_agency

pa = pd.read_csv(PA_PATH + "agency_power_prepro.csv", sep=',')

# get pa per sentence

len(df[df.verb.apply(lambda x: True if type(x) == tuple else False)])

2220 sentences with a preposition

def get_pa(verb):  
    agency = None
    power = None        
    if type(verb) == tuple:
        # (verb, prep)
        try:
            agency = pa[(pa.lemma == verb[0]['lemma']) & (pa.prep == verb[1]['lemma'])].agency.values[0]
        except:
            pass
        try:
            power = pa[(pa.lemma == verb[0]['lemma']) & (pa.prep == verb[1]['lemma'])].power.values[0]
        except:
            pass
    else:
        # just verb
        try:
            agency = pa[(pa.lemma == verb['lemma']) & (~pa.prep.notna())].agency.values[0]
        except:
            pass
        try:
            power = pa[(pa.lemma == verb['lemma']) & (~pa.prep.notna())].power.values[0]
        except:
            pass        
            
    return agency, power

df['agency'], df['power'] = zip(*df.verb.progress_apply(get_pa))

df.head()

len(df)

df = df[(df.agency.notna() & df.power.notna())] # power and agency must be there

len(df)

df.to_pickle(MS_PATH + 'character_sentences_verbs_pa.pkl')

# get genders

%%time
df = pd.read_pickle(MS_PATH + 'character_sentences_verbs_pa.pkl')

## read character metadata

cha = pd.read_csv(MS_PATH + 'character.metadata.tsv', sep='\t', names=['id', 'fid', 'date', 'cha_name', 'actor_birth', 'actor_gender', 'actor_height', 'actor_ethnicity', 'actor_name', 'actor_age', 'cha_actor_fid', 'cha_fid', 'actor_fid'])
cha = cha.drop(columns=['actor_height', 'actor_ethnicity', 'cha_actor_fid'])
cha.index = pd.MultiIndex.from_arrays(cha[['id', 'cha_fid']].values.T, names=['mid', 'cid'])
cha = cha[cha.cha_name.notna() & cha.actor_gender.notna()] 
cha = cha[~cha.index.duplicated('first')] 

def get_gender(cid):
    return cha[cha.cha_fid == cid].actor_gender.values[0]

df['gender'] = df.cid.progress_apply(get_gender)

df.to_pickle(MS_PATH + 'character_sentences_verbs_pa_gender.pkl')

# To target format

%%time
df = pd.read_pickle(MS_PATH + 'character_sentences_verbs_pa_gender.pkl')

- original:  ,sen,sendel,oricat,verbs,storyid,sentencenum 
- here:      ,sen,sendel,agency,power,verbs,mid,senid,charid,gender

data = pd.DataFrame(columns=['sen','sendel','agency','power','verb','mid','senid','charid','gender'])

data['sen'] = df.sentence.progress_apply(sentence_to_raw_string)

MASK = '<VERB>'

def mask_verb(row):
    t_verb = row.verb
    if type(t_verb) == tuple:
        # t_verb[1] is then a preposition
        t_verb = t_verb[0]
    
    tokens = []
    for t in row.sentence['tokens']['token']:
        if t == t_verb:
            tokens.append(MASK)
        elif not t['word'] in ['-LRB-', '-LCB-', '-RCB-', '-RRB-']: # remove special characters
            tokens.append(t['word'])
    
    return ' '.join(tokens)

data['sendel'] = df.progress_apply(mask_verb, axis=1)

data['agency'] = df.agency

data['power'] = df.power

def get_verb(verb):
    t_verb = verb
    if type(t_verb) == tuple:
        # t_verb[1] is then a preposition
        t_verb = t_verb[0]
    return t_verb['lemma']

data['verb'] = df.verb.progress_apply(get_verb)

data['mid'] = [idx[1] for idx in list(df.index)]

data['senid'] = df.sentence.progress_apply(lambda s:s['@id'])

data['charid'] = [idx[0] for idx in list(df.index)]

data['gender'] = df.gender

data.head()

data.tail()

data.gender.value_counts()

data = data.reset_index(drop=True)

data.head()

data.to_csv(MS_PATH + 'pa-transformer/movie_sentences.csv')

## map agency and power

data = pd.read_csv(MS_PATH + 'pa-transformer/movie_sentences.csv', index_col=0)

data.head()

data.agency = data.agency.map({'agency_pos': 'pos', 'agency_equal': 'equal', 'agency_neg': 'neg' })

data.power = data.power.map({'power_agent': 'pos', 'power_equal': 'equal', 'power_theme': 'neg' })

data.agency.value_counts()

data.power.value_counts()

MS_PATH + 'pa-transformer/movie_sentences.csv'

data.to_csv(MS_PATH + 'pa-transformer/movie_sentences.csv')

# filter too long sentences

data = pd.read_csv(MS_PATH + 'pa-transformer/movie_sentences.csv', index_col=0)

data.head()

data.sen.apply(lambda s: len(s.split(' '))).describe()

64 / 2 -4

max_sen_len = 28

test = data[data.sen.apply(lambda s: len(s.split(' ')) <= max_sen_len)]

testf = test[test.gender == 'F'].sample(25000, random_state=42)

testm = test[test.gender == 'M'].sample(25000, random_state=42)

print(testf.agency.value_counts().pos / len(testf))
print(testf.agency.value_counts().equal / len(testf))
print(testf.agency.value_counts().neg / len(testf))

print(testf.power.value_counts().pos / len(testf))
print(testf.power.value_counts().equal / len(testf))
print(testf.power.value_counts().neg / len(testf))

print(testm.agency.value_counts().pos / len(testm))
print(testm.agency.value_counts().equal / len(testm))
print(testm.agency.value_counts().neg / len(testm))

print(testm.power.value_counts().pos / len(testm))
print(testm.power.value_counts().equal / len(testm))
print(testm.power.value_counts().neg / len(testm))

final_data = testf.append(testm)

final_data

final_data.to_csv(MS_PATH + 'pa-transformer/movie_sentences_genderbalanced.csv')
