==> rewriting/data_creation/movie/4_train_dev_test_split.ipynb <==
import os
# nvidia-smi
os.environ['CUDA_VISIBLE_DEVICES'] = '0'

import pandas as pd
import csv
import numpy as np
from tqdm import tqdm
tqdm.pandas()

import warnings
warnings.filterwarnings("ignore")

PREFIX = "../data/"
PA_PATH = PREFIX + "sap2017-connotation-frames-power-agency/"
MS_PATH = PREFIX + "bamman2013-movie-summaries/MovieSummaries/"
NLP_PATH = MS_PATH + 'corenlp_plot_summaries/'

# read data

data = pd.read_csv(MS_PATH + 'pa-transformer/movie_sentences_genderbalanced.csv', index_col = 0)

data.head()

len(data)

(50000*0.7), (50000*0.2), (50000*0.1)

# split

test = data.sample(5000)

test.head()

data = data[~data.isin(test)].dropna()

dev = data.sample(10000)

train = data[~data.isin(dev)].dropna()

len(train), len(dev), len(test)

train.to_csv(MS_PATH + 'pa-transformer/movie_train.csv')

dev.to_csv(MS_PATH + 'pa-transformer/movie_dev.csv')

test.to_csv(MS_PATH + 'pa-transformer/movie_test.csv')

train.gender.value_counts()

dev.gender.value_counts()

test.gender.value_counts()

# view data

train = pd.read_csv(MS_PATH + 'pa-transformer/movie_train.csv')

train.head()

train.agency.value_counts()

train.power.value_counts()

test = pd.read_csv(MS_PATH + 'pa-transformer/movie_test.csv')

len(test)

test.agency.value_counts()

test.power.value_counts()
